Add Handler.{Wait,Close}
This commit adds these two methods to allow the user to wait until all
Handler's background jobs are done to be able to clean up its
dependencies.

This is mostly useful in things that need to be terminated as soon as an
interaction is served, such as in serverless functions.